Add files via upload
There are two files, the file named "pf_a_star.cpp" was the file uploaded for the API purpose having only the final function. The "path_finding_a_star.cpp" file is an independent file with a solution to various test cases in detail along with debugging functionalities for future code rectification.